rebased and refactored on new pebbles master
