+++ pebbles/drivers/provisioning/openshift_driver_config.py
