+++ pebbles/drivers/provisioning/setup.py
@@ -34,6 +34,7 @@ setup(
             'PvcCmdLineDriver = pebbles.drivers.provisioning.pvc_cmdline_driver:PvcCmdLineDriver',
             'OpenStackDriver = pebbles.drivers.provisioning.openstack_driver:OpenStackDriver',
             'DockerDriver = pebbles.drivers.provisioning.docker_driver:DockerDriver',
+            'OpenShiftDriver = pebbles.drivers.provisioning.openshift_driver:OpenShiftDriver',
         ],
     },
